vscode find/replace with regex "for \(int (\w+) = 0; (\w+) < ((?:\w|\.|(?:->))+)\.size\(\); (\w+)\+\+\)" -> "for (size_t $1 = 0; $2 < $3.size(); $4++)"

+++ src/analyticfunctions/FiniteNucleusSphere.h
@@ -37,7 +37,7 @@ public:
 
     double evalf(const mrcpp::Coord<3> &r) const override {
         double result = 0.0;
-        for (int i = 0; i < this->nuclei.size(); i++) {
+        for (size_t i = 0; i < this->nuclei.size(); i++) {
             const auto &R = this->nuclei[i].getCoord();
             auto R1 = math_utils::calc_distance(r, R);
             auto Z = this->nuclei[i].getCharge();
