+++ src/pseudopotential/PPNucleus.h
@@ -66,7 +66,7 @@ public:
     double evalf(const mrcpp::Coord<3> &r) const override {
         double result = 0.0;
         double temp_exp;
-        for (int i = 0; i < this->nuclei.size(); i++) {
+        for (size_t i = 0; i < this->nuclei.size(); i++) {
             const auto &R = this->nuclei[i].getCoord();
             auto R1 = math_utils::calc_distance(R, r);
